+++ changelog.md
@@ -7,3 +7,8 @@ All notable changes to this project will be documented in this file.
 - Added flat increase of tool speed by 50%
 - Added game rule to modify tool speed.
 
+## [1.1.0] - 2/2/2022
+### Added
+- Modded tools should work most of the time.
+- Speed now applies to non tools (Ex: your fist)
+